fix for class loading race condition / deadlock.
the JVM can deadlock if multiple threads try to load classes that
contain references to each other.  this is asinine, and i would argue
a bug in the JVM, but what can you do?  so now we don't reflect in
the UsageMonitor until we are sure the class is already loaded.

a decent article on the issue:
https://blogs.oracle.com/jrockit/entry/class_loading_deadlocks_1